Sync "spawn" command for all entities and restrict "sub" and "spawn" to at least moderator perms
