+++ NitroxClient/GameLogic/NitroxConsole.cs
@@ -14,12 +14,12 @@ namespace NitroxClient.GameLogic
         public static bool DisableConsole { get; set; } = true;
 
         private readonly IPacketSender packetSender;
-        private readonly Items item;
+        private readonly Items items;
 
-        public NitroxConsole(IPacketSender packetSender, Items item)
+        public NitroxConsole(IPacketSender packetSender, Items items)
         {
             this.packetSender = packetSender;
-            this.item = item;
+            this.items = items;
         }
 
         //List of things that can be spawned : https://subnauticacommands.com/items
@@ -35,8 +35,7 @@ namespace NitroxClient.GameLogic
                 }
                 else
                 {
-                    SpawnItem(gameObject);
-                    //TODO: Add support for no AI creature that need to be spawned as well
+                    DefaultSpawn(gameObject);
                 }
             }
             catch (Exception ex)
@@ -46,29 +45,22 @@ namespace NitroxClient.GameLogic
         }
 
         /// <summary>
-        /// Spawns Seamoth, Exosuit and Cyclops
+        /// Spawns Seamoth, Exosuit or Cyclops
         /// </summary>
         private void SpawnVehicle(GameObject gameObject, TechType techType)
         {
             NitroxId id = NitroxEntity.GetIdOrGenerateNew(gameObject);
 
-            VehicleWorldEntity vehicleEntity = Vehicles.MakeVehicleEntity(gameObject, id, techType);
+            VehicleWorldEntity vehicleEntity = Vehicles.BuildVehicleWorldEntity(gameObject, id, techType);
             
             packetSender.Send(new EntitySpawnedByClient(vehicleEntity));
 
             Log.Debug($"Spawning vehicle {techType} with id {id} at {gameObject.transform.position}");
         }
 
-        /// <summary>
-        /// Spawns a Pickupable item
-        /// </summary>
-        private void SpawnItem(GameObject gameObject)
+        private void DefaultSpawn(GameObject gameObject)
         {
-            if (gameObject.TryGetComponent(out Pickupable pickupable))
-            {
-                Log.Debug($"Spawning item {pickupable.GetTechName()} at {gameObject.transform.position}");
-                item.Dropped(gameObject, pickupable.GetTechType());
-            }
+            items.Dropped(gameObject);
         }
 
         private static TechType GetObjectTechType(GameObject gameObject)
